re-enable stop command for playbackstart event
... but this time at a more suitable position. Background:
If an EmbyCon video is playing and you switch to a live tv channel in Kodi without stopping the video playback first, then Kodi for whatever reason sends a "onPlayBackResumed" event.
This does not happen if you start a new video, then it will be "onPlayBackStopped", which makes sense.
But because of "onPlayBackResumed" the EmbyCon stop command will be send at a later time, when the user stops live tv in Kodi.
This can lead to strange effects in Emby if something depends on playback events there.